requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile requirements.in --output-file requirements.txt
aiofiles==24.1.0
    # via -r requirements.in
annotated-types==0.7.0
    # via pydantic
anyio==4.9.0
    # via
    #   google-genai
    #   httpx
    #   starlette
cachetools==5.5.2
    # via google-auth
certifi==2025.7.14
    # via
    #   httpcore
    #   httpx
    #   requests
charset-normalizer==3.4.2
    # via requests
click==8.2.1
    # via uvicorn
fastapi==0.116.1
    # via -r requirements.in
filetype==1.2.0
    # via langchain-google-genai
google-ai-generativelanguage==0.6.18
    # via langchain-google-genai
google-api-core==2.25.1
    # via google-ai-generativelanguage
google-auth==2.40.3
    # via
    #   google-ai-generativelanguage
    #   google-api-core
    #   google-genai
google-genai==1.25.0
    # via -r requirements.in
googleapis-common-protos==1.70.0
    # via
    #   google-api-core
    #   grpcio-status
grpcio==1.73.1
    # via
    #   google-api-core
    #   grpcio-status
    #   qdrant-client
grpcio-status==1.73.1
    # via google-api-core
h11==0.16.0
    # via
    #   httpcore
    #   uvicorn
h2==4.2.0
    # via httpx
hpack==4.1.0
    # via h2
httpcore==1.0.9
    # via httpx
httpx==0.28.1
    # via
    #   -r requirements.in
    #   google-genai
    #   langgraph-sdk
    #   langsmith
    #   qdrant-client
hyperframe==6.1.0
    # via h2
idna==3.10
    # via
    #   anyio
    #   httpx
    #   requests
jsonpatch==1.33
    # via langchain-core
jsonpointer==3.0.0
    # via jsonpatch
langchain==0.3.26
    # via -r requirements.in
langchain-core==0.3.69
    # via
    #   langchain
    #   langchain-google-genai
    #   langchain-text-splitters
    #   langgraph
    #   langgraph-checkpoint
    #   langgraph-prebuilt
langchain-google-genai==2.1.8
    # via -r requirements.in
langchain-text-splitters==0.3.8
    # via langchain
langgraph==0.5.3
    # via -r requirements.in
langgraph-checkpoint==2.1.0
    # via
    #   langgraph
    #   langgraph-prebuilt
langgraph-prebuilt==0.5.2
    # via langgraph
langgraph-sdk==0.1.73
    # via langgraph
langsmith==0.4.6
    # via
    #   langchain
    #   langchain-core
numpy==2.3.1
    # via qdrant-client
orjson==3.11.0
    # via
    #   langgraph-sdk
    #   langsmith
ormsgpack==1.10.0
    # via langgraph-checkpoint
packaging==25.0
    # via
    #   langchain-core
    #   langsmith
portalocker==2.10.1
    # via qdrant-client
proto-plus==1.26.1
    # via
    #   google-ai-generativelanguage
    #   google-api-core
protobuf==6.31.1
    # via
    #   google-ai-generativelanguage
    #   google-api-core
    #   googleapis-common-protos
    #   grpcio-status
    #   proto-plus
    #   qdrant-client
pyasn1==0.6.1
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.4.2
    # via google-auth
pydantic==2.11.7
    # via
    #   fastapi
    #   google-genai
    #   langchain
    #   langchain-core
    #   langchain-google-genai
    #   langgraph
    #   langsmith
    #   pydantic-settings
    #   qdrant-client
pydantic-core==2.33.2
    # via pydantic
pydantic-settings==2.10.1
    # via -r requirements.in
python-dotenv==1.1.1
    # via
    #   -r requirements.in
    #   pydantic-settings
pytimeparse==1.1.8
    # via twelvedata
pyyaml==6.0.2
    # via
    #   langchain
    #   langchain-core
qdrant-client==1.14.3
    # via -r requirements.in
redis==6.2.0
    # via -r requirements.in
requests==2.32.4
    # via
    #   -r requirements.in
    #   google-api-core
    #   google-genai
    #   langchain
    #   langsmith
    #   requests-toolbelt
    #   twelvedata
requests-toolbelt==1.0.0
    # via langsmith
rsa==4.9.1
    # via google-auth
sniffio==1.3.1
    # via anyio
sqlalchemy==2.0.41
    # via langchain
starlette==0.47.1
    # via fastapi
tenacity==8.5.0
    # via
    #   google-genai
    #   langchain-core
tqdm==4.67.1
    # via -r requirements.in
twelvedata==1.2.25
    # via -r requirements.in
typing-extensions==4.14.1
    # via
    #   -r requirements.in
    #   anyio
    #   fastapi
    #   google-genai
    #   langchain-core
    #   pydantic
    #   pydantic-core
    #   sqlalchemy
    #   starlette
    #   typing-inspection
typing-inspection==0.4.1
    # via
    #   pydantic
    #   pydantic-settings
urllib3==2.5.0
    # via
    #   qdrant-client
    #   requests
uvicorn==0.35.0
    # via -r requirements.in
websockets==15.0.1
    # via google-genai
xxhash==3.5.0
    # via langgraph
zstandard==0.23.0
    # via langsmith
